# Fixes #142: Optimized batch norm evaluation